theme and themes directory now under the theme umbrella along with GLobalStyles
