+++ src/App.tsx
@@ -3,7 +3,6 @@ import { Router, Location, RouteComponentProps } from '@reach/router';
 import { ThemeProvider } from 'styled-components';
 import posed, { PoseGroup } from 'react-pose';
 import ReactGA from 'react-ga';
-import GlobalStyles from './GlobalStyles';
 import Header from './ui/Header';
 import Dashboard from './pages/Dashboard';
 import Profile from './pages/Profile';
@@ -17,8 +16,8 @@ import Footer from './ui/Footer';
 import { useInfoButtons, InfoButtonState } from './api/info-buttons';
 import { useUser } from './api/user';
 import { useAppVersions, AppVersions } from './api/app-versions';
-import { themesLookup, defaultTheme } from './themes';
-import { styled } from './theme';
+import { themesLookup, defaultTheme } from './theme/themes';
+import { styled, GlobalStyles } from './theme';
 
 const ContentWrapper = styled.div`
   display: flex;
