+++ src/theme/index.ts
@@ -1,0 +1,16 @@
+import { shadows, gradients, breakpoints, themeSettings } from './theme-settings';
+import { Color } from './theme-colors';
+import { ThemeContext, ThemeConfiguration, styled } from './theme-interface';
+import GlobalStyles from './GlobalStyles';
+export {
+  shadows,
+  gradients,
+  breakpoints,
+  GlobalStyles,
+  themeSettings,
+  Color,
+  ThemeContext,
+  styled
+};
+
+export type ThemeConfiguration = ThemeConfiguration;
