+++ src/theme/theme-colors.ts
@@ -1,0 +1,42 @@
+export enum Color {
+  'transparent' = 'transparent',
+  'white' = '#ffffff',
+  'black' = '#000000',
+  // Neutral Palette
+  'neutral-100' = '#F7F5F5', // only use as an off white background
+  'neutral-200' = '#E9E5E4', // use as a light gray background with dark text
+  'neutral-300' = '#D4CFCD', // use as a gray background with very dark text
+  'neutral-400' = '#B7B1AF', // use carefully
+  'neutral-500' = '#8F8582', // minimum contrast for dark ui (non-text) elements on white - WCAG 2.1
+  'neutral-550' = '#7B746F', // minimum contrast for dark text with white background
+  'neutral-600' = '#696361', // minimum for dark text with background up to neutral-200
+  'neutral-700' = '#423E3C', // use as a text color for backgrounds as dark as neutral-400
+  'neutral-800' = '#2E2B2A',
+  'neutral-900' = '#1F1D1C',
+  // Primary
+  'orange-100' = '#FFE7DE',
+  'orange-200' = '#F79572',
+  'orange-300' = '#E85F2E',
+  'orange-400' = '#D73F09', // Base shade
+  'orange-500' = '#B22F00',
+  'orange-600' = '#7D2100',
+  'orange-700' = '#611A00',
+  // Accents
+  'stratosphere-100' = '#ECF1F4',
+  'stratosphere-200' = '#C6DAE7',
+  'stratosphere-300' = '#2490B5',
+  'stratosphere-400' = '#006A8E', // Base shade
+  'stratosphere-500' = '#005875',
+  'stratosphere-600' = '#003B5C',
+  'stratosphere-700' = '#002235',
+  'pine-100' = '#EDF3E3',
+  'pine-200' = '#C4D6A4',
+  'pine-300' = '#73A663',
+  'pine-400' = '#4A773C', // Base shade
+  'pine-500' = '#3C692E',
+  'pine-600' = '#254A1A',
+  'pine-700' = '#142E0C',
+  'lava-400' = '#C72127',
+  'luminance-300' = '#FFCE57',
+  'roguewave-400' = '#00859B'
+}
