+++ src/theme/theme-settings.ts
@@ -1,0 +1,41 @@
+export const shadows = {
+  1: 'rgba(66, 62, 60, 0.1) 0px 10px 16px, rgba(105, 99, 97, 0.05) 0px 3px 16px',
+  2: '0px 1px 5px 0px rgba(0, 0, 0, 0.2), 0px 2px 2px 0px rgba(0, 0, 0, 0.14), 0px 3px 1px -2px rgba(0, 0, 0, 0.12)',
+  3: '0px 1px 8px 0px rgba(0, 0, 0, 0.2), 0px 3px 4px 0px rgba(0, 0, 0, 0.14), 0px 3px 3px -2px rgba(0, 0, 0, 0.12)',
+  4: '0px 2px 4px -1px rgba(0, 0, 0, 0.2), 0px 4px 5px 0px rgba(0, 0, 0, 0.14), 0px 1px 10px 0px rgba(0, 0, 0, 0.12)'
+};
+
+export const gradients = {
+  1: 'linear-gradient(rgba(0, 0, 0, 0.55), rgba(0, 0, 0, 0.55))'
+};
+
+export const breakpoints = {
+  '768': '768px',
+  '1024': '1024px'
+};
+
+export const themeSettings = {
+  spacing: {
+    unit: 8,
+    mobile: '1rem', // padding/padding or row/column gaps
+    desktop: '3rem'
+  },
+  widths: {
+    contentMax: '1024px'
+  },
+  borderRadius: {
+    '8': '8px',
+    '16': '16px'
+  },
+  fontSize: {
+    '12': '1.2rem',
+    '14': '1.4rem',
+    '15': '1.5rem',
+    '16': '1.6rem',
+    '18': '1.8rem',
+    '20': '2.0rem',
+    '24': '2.4rem',
+    '36': '3.6rem',
+    '58': '5.8rem'
+  }
+};
